Nil keys are no longer equivalent to each other
This avoids the problem of "duplicate" keys which is detected as an error when attempting to assign a new value to the underlying storage, this occurred because as the weak keys become nil the weak key containers became equivalent to each other as their equality was no longer based on the key.

Once an underlying key becomes nil it never needs to be referenced again so this is not a problem.

The same nil key container will still be equivalent when it is the same key container as it compares the hashValues when any of the underlying keys are nil

An alternate fix is to reap null keys prior to insertion but it was two times slower when inserting so I have gone with this solution.